Layout Set
Set out layout and schematics for both LCD and push-buttons. Users will have the option to use either or depending on which solder bridges they connect. Board USB also added to the device, may need to change which side of the board the traces are on if buttons/LCD are not visible.