Reverse USB side
Moved the USB to be on the other side of the board to maintain the sanctity of the zero on the left side (for the imperial units).
Reduced transistor diagrams to not some of the smaller diagrams. Added additional hardware to allow the ATtiny to work with USB.
Added some Mechanical engineering focused constants (Avogadro's Number, Gas Constant, Accl due to gravity in metric & imperial)